moved plugins to plugins.py
